requirements.txt
python-dotenv~=1.1.0
redis~=5.2.1
fastapi~=0.115.12
slowapi~=0.1.9
requests~=2.32.3